+++ docs/change-log.md
@@ -1,6 +1,7 @@
 ### Changes in v0.3 (not released):
 - Added button that locates currently played track
 - Fixed volume and mute state not being synchronized
+- Added "Cursor follows playback" option
 
 ### Changes in v0.2 (released 2018-08-25):
 - First official release